Makefile: Polly: Pass -polly-opt-fusion=max if CLANG ≯ 14.0.0
The command line option -polly-opt-fusion has been completely removed [1] in the latest LLVM toolchain (and in the latest Clang 14 build)

The cmdline option was removed previously [2], though the main switch remained because of a human error
which kept this option functioning in the polly optimizer. But in reality, it did nothing since then.
Now, it has been completely removed.

Let's use a condition to make compiler use this flag if clang's version is less
than 14.

The name of the option is misleading and has been renamed by isl to
"serialize-sccs". Instead of also renaming the option, it has been removed by llvm.

The option is still accessible using the following cmdline option:

[ -polly-isl-arg=--no-schedule-serialize-sccs ]

[1] https://github.com/llvm/llvm-project/commit/cb879d00d8d54383447c1d94ec4ef0b28873dcaa
[2] https://github.com/llvm/llvm-project/commit/07e7cb9433daa80e7a8b418cb334c663abe68202

Change-Id: Ia63924ad013cffa7131e1176131857102de68d63

Signed-off-by: nerd.zenss_ <nerd.zensdev@gmail.com>