{X00TD, X01BD}_defconfig: Enable INIT_STACK_ALL_ZERO
Init stack is secure as well usefull.
As per logs we have certain uninitialised variables.
so this will be a valid fix for them.

Note that as per comments the usage of this is controversial
But, no adverse effects were noticed on our system
So, I feel it's safe to use on our devices

Change-Id: I4abc5e13f373df43201657ed199ab7eb7b0b7930

Signed-off-by: nerd.zenss_ <nerd.zensdev@gmail.com>